Make the example more clear
I was pretty cconfused when reading this doc since the 2nd example is using `WoltLabSuite/Core/Ui/Foo` but `WoltLabSuite/Core/Ui/Foo` doesn't have an actual value.
I also never heard of require.js, so I didn't know anything about the way it works. So It'd sure be better if to refer to the RequireJS-API docs.